feat: add class MotPalabraPoll instead on motPalabraImportance, which wii have a element foreach importance of MotPalabra
